requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
loguru==0.7.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
telebot==0.0.5
urllib3==2.0.3
